requirements.txt
# Framework Web
Flask==3.0.0
Werkzeug==3.0.1

# Segurança
Flask-WTF==1.2.1
Flask-Talisman==1.1.0
Flask-Limiter==3.5.0
python-dotenv==1.0.0

# Validação e Formulários
WTForms==3.1.1
email-validator==2.1.0

# Utilitários
Pillow==10.1.0
